create separate class for network management
